Support for stored procedure with OUTPUT parameter. Support for "execute" as well as "exec".
Added 2 changes

1. Very minor change.  Add support for "execute" as well as "exec" for
those that like to painfully spell everything out  :)

2. Allow support for a SP with an OUTPUT parameter.

The way it's implemented is:
To call a stored procedure with an OUTPUT parameter an additional
parameter must be passed to their edge.func call.

The parameter Key name is "returnParameter" and the parameter value is
the name of the OUTPUT parameter in the SQL procedure (without the @).

The only other requirement is the OUTPUT parameter in the proc must
ultimately a nvarchar(max). Most typical SQL datatypes can be cast to
nvarchar(max) without issue.    nvarchar(max) also passes into a C#
string without issue